+++ .github/workflows/motes-cicd.yml
@@ -14,7 +14,7 @@ jobs:
     name: Job - Miniconda CI on ${{ matrix.os }} with Python ${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}
     strategy:
-      fail-fast: false
+      fail-fast: true
       matrix:
         os: ["ubuntu-latest", "windows-latest"]
         python-version: [3.9]
