+++ conda/motes.yml
@@ -3,6 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
+  - python>=3.9
   - matplotlib
   - astropy
   - scipy
