--- docs/explanation.md
@@ -1,11 +1,0 @@
-# Explanation
-
-Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam sollicitudin hendrerit quam. Proin ex leo, condimentum non sagittis id, vulputate at arcu. Suspendisse aliquet auctor leo scelerisque cursus. Nullam vel fermentum felis, vitae ornare orci. Aliquam mauris felis, pharetra nec risus id, rhoncus vehicula risus. Morbi ac condimentum lectus. Donec turpis felis, iaculis ut sem in, bibendum accumsan urna. Ut in elit eros.
-
-Nunc neque nunc, vulputate non mauris vitae, condimentum rhoncus sapien. Donec euismod nisl non lacinia aliquet. Curabitur vel scelerisque mauris, id efficitur purus. Sed porta, justo sed sodales facilisis, nisl erat imperdiet elit, id dignissim lorem nibh ut massa. Nulla non turpis ut augue eleifend laoreet sit amet id neque. Nunc ultrices velit diam, nec congue mauris maximus nec. Mauris vel neque nec mauris semper vehicula non at ex. Nam lacinia mi eu aliquam rutrum. Vestibulum gravida lectus sit amet enim lobortis, et egestas mauris sodales. Cras sollicitudin quis dui id porttitor. Vestibulum tempor lorem et lorem dignissim mollis. Cras dapibus sapien et sem hendrerit, vitae maximus tortor condimentum. Ut sed eros tristique, rutrum sapien nec, tempor eros.
-
-Sed sit amet neque id arcu tempor tempor. Donec at leo ut leo porta iaculis. Quisque tincidunt nec tellus eu cursus. Sed placerat orci nibh. Nulla sagittis scelerisque urna. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed sit amet neque venenatis, gravida odio at, sodales risus. Donec non elit dolor. Curabitur interdum at est quis ultrices. Integer accumsan commodo tincidunt. In ut ultrices metus, a rhoncus ante.
-
-Nunc aliquam lectus vitae diam consequat luctus. Mauris fermentum arcu sed nisl euismod vehicula. Cras faucibus urna molestie massa cursus consectetur. Aenean mi magna, dictum eu feugiat eu, fringilla eget est. Nulla sem dui, tempor eget sapien vitae, molestie tristique risus. Mauris ornare massa et enim aliquet interdum. Sed lorem justo, imperdiet at est at, ultrices rhoncus eros. Nam rutrum, lacus at venenatis tempor, enim nunc viverra erat, sed porta justo diam sed lectus. Morbi tincidunt turpis non elementum dignissim. Mauris et risus quis elit molestie condimentum eget eget enim. In hac habitasse platea dictumst. Vestibulum fringilla libero nec dolor interdum mattis.
-
-Cras sed mi nisi. Aenean interdum urna dui, quis accumsan ligula facilisis at. Integer quis dui ante. Phasellus tortor sem, ultricies id ligula nec, ultricies porta mauris. Pellentesque convallis ultricies hendrerit. Suspendisse potenti. Nunc commodo convallis pellentesque. In luctus nulla at lectus molestie, scelerisque placerat nulla posuere. Duis dui neque, sodales in enim in, mattis convallis est. Cras sollicitudin vestibulum ex, eget pretium eros sodales id. Curabitur nec mattis urna. Sed pellentesque aliquam libero, tincidunt aliquet felis.
