+++ docs/reference-core.md
@@ -1,0 +1,2 @@
+::: motes.core
+
