--- docs/reference-motes.md
@@ -1,2 +1,0 @@
-::: motes
-
